third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/redirect-cross-origin-tripmine-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/resources/tripmine.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Test that a cross-origin redirect does not result in a non-simple request being sent to the target.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Asynchronous XMLHttpRequest 307 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 307 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 303 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 303 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 302 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 302 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 307 DELETE redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 DELETE redirect:
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 POST redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 GET redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 303 POST redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 303 GET redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=303&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 302 POST redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 302 GET redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=302&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 DELETE redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 POST redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 GET redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 DELETE redirect:
PASS testSyncReq.send('data') threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/redirect.php?code=301&url=http://localhost:8000/resources/tripmine.php'..
PASS tripmineStatus is ""
PASS successfullyParsed is true

TEST COMPLETE
